fix(web): cyrillic РУ short code, move language dialog to profile page
